feat: 가이드북 리뷰 생성 응답 반환 및 rating 타입 int 변경
- GuidebookReviewResponse DTO 및 GuidebookReviewMapper 추가
- rating 컬럼 타입을 float에서 int로 변경 (마이그레이션 포함)
- 리뷰 생성 시 응답 반환하도록 서비스 수정
- 서비스 테스트를 mock 대신 실제 객체(픽스처 메서드) 기반으로 리팩토링

+++ src/main/java/region/jidogam/domain/guidebook/entity/GuidebookReview.java
@@ -38,6 +38,5 @@ public class GuidebookReview extends BaseUpdatableEntity {
 
   @Column(nullable = false)
   @Builder.Default
-  private Double rating = 0.0;
-
+  private Integer rating = 0;
 }
